// src/app/modules/Admin/Admin.interface.ts
import { Model } from "mongoose";

export type TGender = 'male' | 'female' | 'other';

export type TBloodGroup =
    | 'A+'
    | 'A-'
    | 'B+'
    | 'B-'
    | 'AB+'
    | 'AB-'
    | 'O+'
    | 'O-';

export type TUserName = {
    firstName: string;
    middleName: string;
    lastName: string;
}

export type TAdmin = {
    id: string;
    user: string;
    designation: string;
    name: TUserName;
    gender: TGender;
    dateOfBirth?: Date;
    email: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup?: TBloodGroup;
    presentAddress: string;
    permanentAddress: string;
    profileImg?: string;
    isDeleted: boolean;
}

export interface AdminModel extends Model<TAdmin> {
   
} 